requirements_python.txt
paho-mqtt
numpy
pymysql
psycopg2-binary
sqlalchemy
pandas
